modularize DeterministicProfileClient into separate files, fix lint errors
